Change user role to investor when registering as an investor
