+++ lib/db.ts
@@ -1,7 +1,12 @@
 import { drizzle } from "drizzle-orm/neon-http";
 import { InvestorFormData, InvestorRequestData } from "../types/index";
-import { InvestorTable, InvestorRequestTable } from "./schema";
-import {neon} from "@neondatabase/serverless";
+import { 
+  InvestorTable, 
+  InvestorRequestTable,
+  UserTable,
+} from "./schema";
+import { eq } from "drizzle-orm";
+import { neon } from "@neondatabase/serverless";
 import dotenv from 'dotenv';
 import path from 'path';
 
@@ -15,27 +20,4 @@ if (!databaseUrl) {
 
 const sql = neon(databaseUrl);
 const db = drizzle(sql);
-
-export async function addInvestor(investor: InvestorFormData) {
-  console.log("Investor Data before submission:", investor);
-
-  const investorData = {
-    ...investor,
-    birthDate: investor.birthDate ? new Date(investor.birthDate).toISOString() : '',
-  };
-
-  const insertedInvestor = await db
-    .insert(InvestorTable)
-    .values(investorData)
-    .returning({ investorId: InvestorTable.id })
-    .execute();
-
-  return insertedInvestor[0]?.investorId;
-}
-
-export async function addInvestorRequest(request: InvestorRequestData) {
-  return await db.insert(InvestorRequestTable).values(request).execute();
-}
-
-
